back-end/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
channels==4.0.0  